Refactor RegionalScores to follow standard react patterns
